# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para gestão integral da produção gráfica.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Pesquisar ordens de produção
      parameters:
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Ordens correspondentes.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrderSummary" }
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Consultar OP e fases de produção
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Detalhe da ordem.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetails" }
        "404": { $ref: "#/components/responses/NotFound" }
  /orders/{orderId}/stages:
    post:
      operationId: updateOrderStage
      tags: [orders]
      summary: Iniciar ou concluir uma etapa produtiva
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderStageInput" }
      responses:
        "200":
          description: Etapa registada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetails" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /orders/{orderId}/quality-check:
    post:
      operationId: submitQualityCheck
      tags: [orders]
      summary: Registar controlo de qualidade
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QualityCheckInput" }
      responses:
        "200":
          description: Controlo de qualidade registado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetails" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    BadRequest:
      description: Dados inválidos ou transição não permitida.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Registo não encontrado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    TimelineEvent:
      type: object
      required: [id, ticketId, ticketCode, orderId, actorName, action, previousState, newState, note, createdAt]
      properties:
        id: { type: integer }
        ticketId: { type: integer }
        ticketCode: { type: string }
        orderId: { type: ["integer", "null"] }
        actorName: { type: string }
        action: { type: string }
        previousState: { type: [string, "null"] }
        newState: { type: [string, "null"] }
        note: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
...
    OrderSummary:
      type: object
      required: [id, orderCode, ticketId, ticketCode, clientName, product, quantity, status, currentStage, dueDate, approvedVersion, createdAt, updatedAt]
      properties:
        id: { type: integer }
        orderCode: { type: string }
        ticketId: { type: integer }
        ticketCode: { type: string }
        clientName: { type: string }
        product: { type: string }
        quantity: { type: integer }
        status: { type: string }
        currentStage: { type: string }
        dueDate: { type: string, format: date }
        approvedVersion: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrderStage:
      type: object
      required: [id, stage, status, responsible, machine, startedAt, completedAt, quantityProduced, waste, notes]
      properties:
        id: { type: integer }
        stage: { type: string }
        status: { type: string }
        responsible: { type: [string, "null"] }
        machine: { type: [string, "null"] }
        startedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        quantityProduced: { type: ["integer", "null"] }
        waste: { type: ["integer", "null"] }
        notes: { type: [string, "null"] }
    QualityCheck:
      type: object
      required: [id, checklist, decision, notes, checkedBy, checkedAt]
      properties:
        id: { type: integer }
        checklist:
          type: object
          required: [quantity, format, material, colors, finishing, approvedArtwork, defectFree, packaging]
          properties:
            quantity: { type: boolean }
            format: { type: boolean }
            material: { type: boolean }
            colors: { type: boolean }
            finishing: { type: boolean }
            approvedArtwork: { type: boolean }
            defectFree: { type: boolean }
            packaging: { type: boolean }
        decision: { type: string, enum: [APROVADO, REPROCESSAR] }
        notes: { type: [string, "null"] }
        checkedBy: { type: string }
        checkedAt: { type: string, format: date-time }
    Delivery:
      type: object
      required: [method, responsible, dispatchAt, deliveredAt, proof, signature, photoPath]
      properties:
        method: { type: [string, "null"], enum: [LEVANTAMENTO, ENTREGA_PROPRIA, TRANSPORTADORA, null] }
        responsible: { type: [string, "null"] }
        dispatchAt: { type: ["string", "null"], format: date-time }
        deliveredAt: { type: ["string", "null"], format: date-time }
        proof: { type: [string, "null"] }
        signature: { type: [string, "null"] }
        photoPath: { type: [string, "null"] }
    OrderDetails:
      allOf:
        - { $ref: "#/components/schemas/OrderSummary" }
        - type: object
          required: [approvedFileName, approvedObjectPath, stages, qualityCheck, delivery, events]
          properties:
            approvedFileName: { type: string }
            approvedObjectPath: { type: string }
            stages:
              type: array
              items: { $ref: "#/components/schemas/OrderStage" }
            qualityCheck:
              oneOf:
                - { $ref: "#/components/schemas/QualityCheck" }
                - { type: "null" }
            delivery: { $ref: "#/components/schemas/Delivery" }
            events:
              type: array
              items: { $ref: "#/components/schemas/TimelineEvent" }
    OrderStageInput:
      type: object
      required: [stage, action]
      properties:
        stage: { type: string, enum: [PRE_IMPRESSAO, IMPRESSAO, ACABAMENTO, EMBALAGEM] }
        action: { type: string, enum: [INICIAR, CONCLUIR] }
        responsible: { type: [string, "null"] }
        machine: { type: [string, "null"] }
        quantityProduced: { type: ["integer", "null"], minimum: 0 }
        waste: { type: ["integer", "null"], minimum: 0 }
        notes: { type: [string, "null"] }
    QualityCheckInput:
      type: object
      required: [checklist, decision]
      properties:
        checklist:
          type: object
          required: [quantity, format, material, colors, finishing, approvedArtwork, defectFree, packaging]
          properties:
            quantity: { type: boolean }
            format: { type: boolean }
            material: { type: boolean }
            colors: { type: boolean }
            finishing: { type: boolean }
            approvedArtwork: { type: boolean }
            defectFree: { type: boolean }
            packaging: { type: boolean }
        decision: { type: string, enum: [APROVADO, REPROCESSAR] }
        notes: { type: [string, "null"] }
...