# lib/api-zod/src/generated/types/qualityCheckInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API para gestão integral da produção gráfica.
 * OpenAPI spec version: 0.1.0
 */
import type { QualityCheckInputChecklist } from './qualityCheckInputChecklist';
import type { QualityCheckInputDecision } from './qualityCheckInputDecision';

export interface QualityCheckInput {
  checklist: QualityCheckInputChecklist;
  decision: QualityCheckInputDecision;
  /** @nullable */
  notes?: string | null;
}