# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API para gestão integral da produção gráfica.
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: number;
  ticketId: number;
  ticketCode: string;
  /** @nullable */
  orderId: number | null;
  actorName: string;
  action: string;
  /** @nullable */
  previousState: string | null;
  /** @nullable */
  newState: string | null;
  /** @nullable */
  note: string | null;
  createdAt: Date;
}